buildings map minor refactoring

+++ Turret_1/src/map_structures/buildings/buildings_map/buildings_map.h
@@ -25,7 +25,7 @@ private:
 	std::vector<std::shared_ptr<Building>> cores;
 	std::vector<TileCoord> justChangedTiles;
 	
-	void createAuxilary(const short size, const TileCoord mainTile, Team* team);
+	void createAuxilary(const uint8_t size, const TileCoord mainTile, Team* team);
 	void pushChanges();
 
 public:
